odb: Refine dbSet size checks for performance and clarity
As suggested in code review:
- Simplify dbSet<T>::hasMoreThan and hasExactly to decrement the argument
  directly, avoiding an extra internal counter.
- Cache getITerms() and getBTerms() into local variables in TritonRoute
  and GlobalRouter to avoid redundant calls.
- Cache iterm->getNet() into a local variable in Rebuffer.cc.
- Add static_cast<uint32_t> to kMaxNetFanout in request_handler.cpp
  to prevent sign-conversion warnings when passed to hasMoreThan.

+++ src/web/src/request_handler.cpp
@@ -2023,7 +2023,7 @@ WebSocketResponse SelectHandler::handleSchematicCone(
             }
             odb::dbNet* net = iterm->getNet();
             if (!net || seen_nets.contains(net)
-                || net->getITerms().hasMoreThan(kMaxNetFanout)) {
+                || net->getITerms().hasMoreThan(static_cast<uint32_t>(kMaxNetFanout))) {
               continue;
             }
             seen_nets.insert(net);
@@ -2065,7 +2065,7 @@ WebSocketResponse SelectHandler::handleSchematicCone(
             }
             odb::dbNet* net = iterm->getNet();
             if (!net || seen_nets.contains(net)
-                || net->getITerms().hasMoreThan(kMaxNetFanout)) {
+                || net->getITerms().hasMoreThan(static_cast<uint32_t>(kMaxNetFanout))) {
               continue;
             }
             seen_nets.insert(net);
